requirements.txt
aiogram
asyncio
requests
todoist-python
aioschedule
bs4
selenium